+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 0.3.3
 -----
     - Added date_izolator(). Fixed bugs in comments parsing with relative dates.
+    - Fixed bug in parsing of Blogpost's content.
 
 0.3.2
 -----